requirements.txt
ansitable==0.9.7
cbor==1.0.0
colored==1.4.3
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.4
packaging==21.3
pgraph-python==0.6.1
Pillow==9.2.0
progress==1.6
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==24.0.1
roboticstoolbox-python==1.0.2
rtb-data==1.0.0
scipy==1.9.3
six==1.16.0
spatialgeometry==1.0.1
spatialmath-python==1.0.1
swift-sim==1.0.0
websockets==10.3
